+++ src/main/java/calculator/DivideOperator.java
@@ -1,7 +1,7 @@
 package main.java.calculator;
 
 public class DivideOperator {
-    public double operate(double num1, double num2) {
+    public int operate(int num1, int num2) {
         if (num2 == 0) {
             throw new ArithmeticException("Divide by zero");
         }
