+++ src/main/java/calculator/MultiplyOperator.java
@@ -1,7 +1,7 @@
 package main.java.calculator;
 
 public class MultiplyOperator {
-    public double operate(double num1, double num2) {
+    public int operate(int num1, int num2) {
         return num1 * num2;
     }
 }
